Assignment 9: Associative Container Map
Problem Statement: 
Implement a map associative container in which the keys will be the Permanent Registration Number (PRN) of the students and the values will be the name of the respective student. Develop a program that prompts the user for PRN and then it looks in the map, using the PRN as an index and returns the name of the student.